+++ admin/src/scenes/utilisateur/list.jsx
@@ -8,7 +8,7 @@ import dayjs from "@/utils/dayjs.utils";
 import { Badge, Container, DropdownButton, Header, Page } from "@snu/ds/admin";
 import { BsDownload } from "react-icons/bs";
 import { IoFlashOutline } from "react-icons/io5";
-import { canViewReferent, formatLongDateFR, getDepartmentNumber } from "snu-lib";
+import { canViewReferent, formatLongDateFR, getDepartmentNumber, canSigninAs } from "snu-lib";
 import Loader from "../../components/Loader";
 import { ExportComponent, Filters, ResultTable, Save, SelectedFilters, SortOption } from "../../components/filters-system-v2";
 import ModalChangeTutor from "../../components/modals/ModalChangeTutor";
@@ -318,13 +318,6 @@ const Action = ({ hit, structure }) => {
     });
   };
 
-  const canTakePlace = () => {
-    if (user.role === ROLES.ADMIN) return true;
-    if (user.role === ROLES.REFERENT_DEPARTMENT && user.department.includes(hit.department) && [ROLES.REFERENT_CLASSE, ROLES.ADMINISTRATEUR_CLE].includes(hit.role)) return true;
-    if (user.role === ROLES.REFERENT_REGION && user.region === hit.region && [ROLES.REFERENT_CLASSE, ROLES.ADMINISTRATEUR_CLE].includes(hit.role)) return true;
-    return false;
-  };
-
   const onConfirmDelete = async () => {
     try {
       const { ok, code } = await api.remove(`/referent/${hit._id}`);
@@ -338,6 +331,7 @@ const Action = ({ hit, structure }) => {
       return toastr.error("Oups, une erreur est survenue pendant la supression du profil :", translate(e.code));
     }
   };
+
   return (
     <>
       <DropdownButton
@@ -359,7 +353,7 @@ const Action = ({ hit, structure }) => {
                   </Link>
                 ),
               },
-              canTakePlace() ? { key: "takePlace", render: <p>Prendre sa place</p>, action: handleImpersonate } : null,
+              canSigninAs(user, hit, "referent") ? { key: "takePlace", render: <p>Prendre sa place</p>, action: handleImpersonate } : null,
               canDeleteReferent({ actor: user, originalTarget: hit, structure }) ? { key: "delete", render: <p>Supprimer le profil</p>, action: onClickDelete } : null,
             ].filter(Boolean),
           },