+++ api/src/controllers/gouv.fr/index.js
@@ -1,0 +1,6 @@
+const express = require("express");
+const router = express.Router();
+
+router.use("/api-education", require("./api-education"));
+
+module.exports = router;